+++ src/main/groovy/com/avioconsulting/mule/testing/dsl/invokers/BatchRunner.groovy
@@ -1,0 +1,10 @@
+package com.avioconsulting.mule.testing.dsl.invokers
+
+// main purpose is to only offer a subset of FlowRunner's methods
+interface BatchRunner {
+    def java(@DelegatesTo(JavaInvoker) Closure closure)
+
+    def json(@DelegatesTo(JsonInvoker) Closure closure)
+
+    def withInputEvent(Closure closure)
+}